+++ src/lib/utils/__tests__/utils.test.ts
@@ -4,8 +4,10 @@ import {
 	calculateNumColumns,
 	isColorGroups,
 	isColorValid,
+	isSameColor,
 	normalizeColor,
 	normalizeInputType,
+	parseColor,
 	transformColors,
 } from '../utils.js'
 
@@ -110,6 +112,12 @@ describe('utils', () => {
 			[colors, { ...params, isCompact: true, compactColorIndices: null }, []],
 			[colors, { ...params, allowDuplicates: false }, colorsObjects],
 			[[1, 1, 1], { ...params, allowDuplicates: false }, [{ value: 1 }]],
+			[['#FF0000', '#ff0000'], { ...params, allowDuplicates: false }, [{ value: '#FF0000' }]],
+			[
+				['#FF0000', '#ff0000'],
+				{ ...params, allowDuplicates: true },
+				[{ value: '#FF0000' }, { value: '#ff0000' }],
+			],
 			[[1, 1, 1], { ...params, allowDuplicates: true }, [{ value: 1 }, { value: 1 }, { value: 1 }]],
 			[colors, { ...params, maxColors: 8 }, colorsObjects],
 			[colors, { ...params, maxColors: 3 }, colorsObjects.slice(0, 3)],
@@ -219,14 +227,75 @@ describe('utils', () => {
 		})
 	})
 
+	describe('parseColor', () => {
+		test.each([
+			['#ff0000', { r: 255, g: 0, b: 0, a: 1 }],
+			['ff0', { r: 255, g: 255, b: 0, a: 1 }],
+			['#f00a', { r: 255, g: 0, b: 0, a: 170 / 255 }],
+			['#ff000080', { r: 255, g: 0, b: 0, a: 128 / 255 }],
+			['rgb(255, 0, 0)', { r: 255, g: 0, b: 0, a: 1 }],
+			['rgb(255 0 0)', { r: 255, g: 0, b: 0, a: 1 }],
+			['rgb(100% 0% 0%)', { r: 255, g: 0, b: 0, a: 1 }],
+			['rgba(0, 128, 0, 0.5)', { r: 0, g: 128, b: 0, a: 0.5 }],
+			['rgb(255 0 0 / 50%)', { r: 255, g: 0, b: 0, a: 0.5 }],
+			['hsl(0 100% 50%)', { r: 255, g: 0, b: 0, a: 1 }],
+			['hsl(120, 100%, 50%)', { r: 0, g: 255, b: 0, a: 1 }],
+			['hsla(240, 100%, 50%, 0.5)', { r: 0, g: 0, b: 255, a: 0.5 }],
+			['hsl(0 0% 50%)', { r: 128, g: 128, b: 128, a: 1 }],
+			['hsl(0rad 100% 50%)', { r: 255, g: 0, b: 0, a: 1 }],
+			['hsl(100grad 100% 50%)', { r: 128, g: 255, b: 0, a: 1 }],
+			['hsl(0.25turn 100% 50%)', { r: 128, g: 255, b: 0, a: 1 }],
+			['red', { r: 255, g: 0, b: 0, a: 1 }],
+			['RebeccaPurple', { r: 102, g: 51, b: 153, a: 1 }],
+			['not-a-color', null],
+			['rgb(0, 0)', null],
+			['rgb(255,,0,0)', null],
+			['rgb(255, 0 0)', null],
+			['rgb(255 0 0 0)', null],
+			['hsl(0 100% 50% 0.5)', null],
+			['rgb(100 %, 0%, 0%)', null],
+			['hsl(0, 100, 50)', null],
+			['constructor', null],
+			['__proto__', null],
+			['toString', null],
+			['hasOwnProperty', null],
+			['valueOf', null],
+			[null, null],
+			[undefined, null],
+			[123, null],
+		])('color:%j, expected:%j', (color, expected) => {
+			expect(parseColor(color)).toEqual(expected)
+		})
+	})
+
 	describe('normalizeColor', () => {
 		test.each([
 			['#ff0', '#ff0'],
 			['#ff0000', '#ff0000'],
+			['#f00a', '#f00a'],
+			['#ff000080', '#ff000080'],
+			['#FF0000', '#ff0000'],
+			['#ABCDEF', '#abcdef'],
+			[' #FfF ', '#fff'],
 			['rgb(255, 0, 0)', '#ff0000'],
+			['rgb(255 0 0)', '#ff0000'],
+			['rgb(100% 0% 0%)', '#ff0000'],
 			['rgba(255, 0, 0, 1)', '#ff0000'],
 			['rgb(0,128,0)', '#008000'],
+			['rgba(255, 0, 0, 0.5)', '#ff000080'],
+			['rgb(255 0 0 / 50%)', '#ff000080'],
+			['hsl(0 100% 50%)', '#ff0000'],
+			['hsl(120, 100%, 50%)', '#00ff00'],
+			['hsla(240, 100%, 50%, 0.5)', '#0000ff80'],
+			['red', '#ff0000'],
+			['RebeccaPurple', '#663399'],
+			[' ff0000 ', '#ff0000'],
+			[' #ff0000 ', '#ff0000'],
+			['rgba(0, 0, 0, 0.999)', '#000000'],
+			['rgb(255 0 0 0)', 'rgb(255 0 0 0)'],
 			['not-a-color', 'not-a-color'],
+			['constructor', 'constructor'],
+			['__proto__', '__proto__'],
 		])('color:%j, expected:%j', (color, expected) => {
 			expect(normalizeColor(color)).toBe(expected)
 		})
@@ -250,11 +319,53 @@ describe('utils', () => {
 			['ffg', false],
 			['.ff0', false],
 			['ff0ff000.', false],
+			['red', true],
+			['RED', true],
+			['rebeccapurple', true],
+			['notacolor', false],
+			['rgb(255, 0, 0)', true],
+			['rgb(255 0 0)', true],
+			['rgba(255, 0, 0, 0.5)', true],
+			['rgb(255 0 0 / 50%)', true],
+			['hsl(0 100% 50%)', true],
+			['hsl(0, 100%, 50%)', true],
+			['hsla(0, 100%, 50%, 0.5)', true],
+			['rgb(0, 0)', false],
+			['hsl(0, 100, 50)', false],
+			['rgb(a, b, c)', false],
+			['rgb(255,,0,0)', false],
+			['rgb(255, 0 0)', false],
+			['rgb(255 0 0 0)', false],
+			['hsl(0 100% 50% 0.5)', false],
+			['rgb(255 0 0 /)', false],
+			['rgb(100 %, 0%, 0%)', false],
+			['rgb(255 0 0 / 5 0%)', false],
+			['constructor', false],
+			['__proto__', false],
+			['toString', false],
 			[null, false],
 			[undefined, false],
 			[0, false],
 		])('color:%j, expected:%j', (color, expected) => {
 			expect(isColorValid(color)).toBe(expected)
+		})
+	})
+
+	describe('isSameColor', () => {
+		test.each([
+			['#ff0000', '#ff0000', true],
+			['#FF0000', '#ff0000', true],
+			['#ABCDEF', '#abcdef', true],
+			['RED', 'red', true],
+			['#ff0000', '#00ff00', false],
+			['red', '#ff0000', false],
+			[null, null, true],
+			['#ff0000', null, false],
+			[null, '#ff0000', false],
+			[1, 1, true],
+			[1, 2, false],
+		])('a:%j, b:%j, expected:%j', (a, b, expected) => {
+			expect(isSameColor(a, b)).toBe(expected)
 		})
 	})